requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pyautogui
pygetwindow
Pillow
